Added empty commands and autonomous SendableChooser
Added AUTO, DEFENSE, and POS commands

AUTO commands drive the bot from a defense to the goal to score in the low goal, with the exception of Approach Defense, which drives the bot forward to a defense.
- AUTO Approach Defense
- AUTO 2 to Goal
- AUTO 3 to Goal
- AUTO 4 to Goal
- AUTO 5 to Goal

DEFENSE commands traverse the bot through various defenses
- DEFENSE Portcullis
- DEFENSE CDF
- DEFENSE Moat
- DEFENSE Ramparts
- DEFENSE Drawbridge
- DEFENSE Sally Port
- DEFENSE Rock Wall
- DEFENSE Rough Terrain

POS commands are used solely for the SendableChooser setup, configuring the autonomous command at the time of the game. The commands change a field indicating the position the bot is at to start autonomous.
- POS 2
- POS 3
- POS 4
- POS 5